+++ README.md
@@ -1,2 +1,2 @@
 # Ion
-Ion is a plugin built around Star Legacy's code base, specifically built for the Horizon's End Minecraft Server.
+Ion is a plugin built around Star Legacy's code base
